Move base field types in to a file that can be required by complex fields
